+++ packages/astro/src/content/runtime.ts
@@ -214,13 +214,13 @@ export function createGetEntry({ liveCollections }: { liveCollections: LiveColle
 				data,
 				collection,
 			} as DataEntryResult | ContentEntryResult;
-			// TODO: remove in Astro 7
+			// TODO: remove in Astro 8
 			warnForPropertyAccess(
 				result.data,
 				'slug',
 				`[content] Attempted to access deprecated property on "${collection}" entry.\nThe "slug" property is no longer automatically added to entries. Please use the "id" property instead.`,
 			);
-			// TODO: remove in Astro 7
+			// TODO: remove in Astro 8
 			warnForPropertyAccess(
 				result,
 				'render',