+++ packages/integrations/node/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @astrojs/node
 
+## 11.1.4
+
+### Patch Changes
+
+- Updated dependencies [[`05763a0`](https://github.com/withastro/astro/commit/05763a0884aabb1da78a2749d5bb9d41ae620527)]:
+  - @astrojs/internal-helpers@0.10.4
+
+## 11.1.3
+
+### Patch Changes
+
+- [#17636](https://github.com/withastro/astro/pull/17636) [`51723b1`](https://github.com/withastro/astro/commit/51723b100a37d6dd6df793957d35d9216e872cef) Thanks [@matthewp](https://github.com/matthewp)! - Updates the adapter to wait for the configured log destination through Astro's new `app.getLogger()` API. This release requires Astro 7.2.1 or later.
+
+- Updated dependencies [[`8c193f6`](https://github.com/withastro/astro/commit/8c193f67cce77cf2e41fb702c88ca46f788f1277)]:
+  - @astrojs/internal-helpers@0.10.3
+
 ## 11.1.2
 
 ### Patch Changes